API_Bitrix/API_B/requirements.txt
pandas
openpyxl
psycopg2
asyncpg
python-dotenv 